+++ packages/mobile/src/message/handleMessage.ts
@@ -8,7 +8,7 @@ import { messageHandlers as analytics } from './handlers/analytics'
 import { messageHandlers as android } from './handlers/android'
 import { messageHandlers as audio } from './handlers/audio'
 import { messageHandlers as cast } from './handlers/cast'
-import { messageHandlers as clientStore } from './handlers/clientStore'
+import { messageHandlers as commonState } from './handlers/commonState'
 import { messageHandlers as download } from './handlers/download'
 import { messageHandlers as haptics } from './handlers/haptics'
 import { messageHandlers as lifecycle } from './handlers/lifecycle'
@@ -29,7 +29,7 @@ const messageHandlers: Partial<MessageHandlers> = {
   ...(isIos ? {} : android),
   ...audio,
   ...cast,
-  ...clientStore,
+  ...commonState,
   ...download,
   ...haptics,
   ...lifecycle,
