+++ packages/mobile/src/store/common/slice.ts
@@ -5,7 +5,7 @@ const initialState: any = {}
 // Slice
 
 const slice = createSlice({
-  name: 'CLIENT_STORE',
+  name: 'COMMON',
   initialState,
   reducers: {
     receive: (state, action) => {
